# helm/usdsearch/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "USD Search and Asset Graph Search APIs",
...
    "version": "1.4.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/vlm_validate/search_result": {
      "post": {
        "tags": [
          "Relevance verification"
        ],
        "summary": "Validate a single search result with VLM",
        "description": "Validates a single asset image against a query using a Vision Language Model. Returns match status, confidence, similarity score, and reasoning.",
        "operationId": "vlm_validate_search_result_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateResultRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryRelevanceValidationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "HTTPBasic": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "QueryRelevanceValidationResult": {
        "properties": {
          "is_match": {
            "type": "boolean",
            "title": "Is Match",
            "description": "Whether the result matches the query"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Confidence in the decision (0-1)"
          },
          "similarity_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Similarity Score",
            "description": "Query-asset relevancy score (0-100)"
          },
          "reasoning": {
            "type": "string",
            "title": "Reasoning",
            "description": "Explanation of the validation decision"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "Identifier of the VLM service+model that produced this verdict. Returned so clients can cache verdicts under a model-aware key \u2014 when the server-side model changes, cached entries become misses and re-validation runs."
          }
        },
        "type": "object",
        "required": [
          "is_match",
          "confidence",
          "similarity_score",
          "reasoning"
        ],
        "title": "QueryRelevanceValidationResult",
        "description": "VLM validation result for a search result."
      },
...
      "ValidateResultRequest": {
        "properties": {
          "asset_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Url",
            "description": "Asset URL to validate (mutually exclusive with image_key/image_keys)"
          },
          "image_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Key",
            "description": "Single image key to validate (mutually exclusive with asset_url/image_keys)"
          },
          "image_keys": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Keys",
            "description": "List of image keys for multiple views of the asset (up to 8). Mutually exclusive with asset_url/image_key."
          },
          "query_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Text",
            "description": "Text query to validate against"
          },
          "query_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Image",
            "description": "Base64 image query to validate against"
          }
        },
        "type": "object",
        "title": "ValidateResultRequest",
        "description": "Request body for single-result VLM validation endpoint."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
...
}